Fix curve and time functions

+++ src/mpu/state.rs
@@ -1,6 +1,6 @@
 pub struct Event {
-    pub value: f32,
-    pub duration: f32,
+    pub value: f64,
+    pub duration: f64,
 }
 
 pub struct MidiState {
@@ -11,8 +11,8 @@ impl MidiState {
     pub fn new() -> MidiState {
         MidiState {
             event: Event {
-                value: 0f32,
-                duration: 0f32,
+                value: 0.0,
+                duration: 0.0,
             },
         }
     }
